+++ example/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pytorch_lite_example/RunModelByCameraDemo.dart';
 import 'package:pytorch_lite_example/RunModelByImageDemo.dart';
 
-void main() => runApp(ChooseDemo());
+void main() => runApp(const ChooseDemo());
 
 class ChooseDemo extends StatefulWidget {
   const ChooseDemo({Key? key}) : super(key: key);
@@ -28,13 +28,13 @@ class _ChooseDemoState extends State<ChooseDemo> {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                          builder: (context) => RunModelByCameraDemo()),
+                          builder: (context) => const RunModelByCameraDemo()),
                     )
                   },
                   style: TextButton.styleFrom(
                     backgroundColor: Colors.blue,
                   ),
-                  child: Text(
+                  child: const Text(
                     "Run Model with Camera",
                     style: TextStyle(
                       color: Colors.white,
@@ -46,13 +46,13 @@ class _ChooseDemoState extends State<ChooseDemo> {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                          builder: (context) => RunModelByImageDemo()),
+                          builder: (context) => const RunModelByImageDemo()),
                     )
                   },
                   style: TextButton.styleFrom(
                     backgroundColor: Colors.blue,
                   ),
-                  child: Text(
+                  child: const Text(
                     "Run Model with Image",
                     style: TextStyle(
                       color: Colors.white,