+++ test/pytorch_lite_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pytorch_lite/pytorch_lite.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('pytorch_lite');